refactor: move model generation to its own file
